Adapt to the latest breaking changes in the firmware, which are make room for our desired current limiting features.

+++ lib/jrk_settings.c
@@ -56,8 +56,8 @@ struct jrk_settings
   uint16_t motor_max_deceleration_reverse;
   uint16_t motor_max_duty_cycle_forward;
   uint16_t motor_max_duty_cycle_reverse;
-  uint8_t motor_max_current_forward;
-  uint8_t motor_max_current_reverse;
+  uint16_t motor_max_current_forward;
+  uint16_t motor_max_current_reverse;
   int8_t motor_current_calibration_forward;
   int8_t motor_current_calibration_reverse;
   uint32_t motor_brake_duration_forward;
@@ -810,25 +810,25 @@ uint16_t jrk_settings_get_motor_max_duty_cycle_reverse(const jrk_settings * sett
   return settings->motor_max_duty_cycle_reverse;
 }
 
-void jrk_settings_set_motor_max_current_forward(jrk_settings * settings, uint8_t motor_max_current_forward)
+void jrk_settings_set_motor_max_current_forward(jrk_settings * settings, uint16_t motor_max_current_forward)
 {
   if (settings == NULL) { return; }
   settings->motor_max_current_forward = motor_max_current_forward;
 }
 
-uint8_t jrk_settings_get_motor_max_current_forward(const jrk_settings * settings)
+uint16_t jrk_settings_get_motor_max_current_forward(const jrk_settings * settings)
 {
   if (settings == NULL) { return 0; }
   return settings->motor_max_current_forward;
 }
 
-void jrk_settings_set_motor_max_current_reverse(jrk_settings * settings, uint8_t motor_max_current_reverse)
+void jrk_settings_set_motor_max_current_reverse(jrk_settings * settings, uint16_t motor_max_current_reverse)
 {
   if (settings == NULL) { return; }
   settings->motor_max_current_reverse = motor_max_current_reverse;
 }
 
-uint8_t jrk_settings_get_motor_max_current_reverse(const jrk_settings * settings)
+uint16_t jrk_settings_get_motor_max_current_reverse(const jrk_settings * settings)
 {
   if (settings == NULL) { return 0; }
   return settings->motor_max_current_reverse;
